Authorization Check Provided

+++ WebAPI/Controllers/ReservationController.cs
@@ -1,11 +1,14 @@
 ﻿using Business.Abstract;
 using Core.Entities;
 using Entities.Reservations;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using WebAPI.Controllers;
 
 [ApiController]
 [Route("api/reservations")]
-public class ReservationController : ControllerBase
+public class ReservationController : BaseController
 {
     private readonly IReservationService _reservationService;
 
@@ -15,6 +18,7 @@ public class ReservationController : ControllerBase
     }
 
     [HttpGet]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin")]
     public IActionResult GetAll()
     {
         var result = _reservationService.GetAll();
@@ -24,17 +28,26 @@ public class ReservationController : ControllerBase
     }
 
     [HttpGet("{id}")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> GetByIdAsync(int id)
     {
         var result = await _reservationService.GetByIdAsync(id);
-        return result.Success
-            ? Ok(result)
-            : BadRequest(result.Message);
+
+        if (!result.Success)
+            return BadRequest(result.Message);
+
+        return result.Data.CustomerId != GetCurrentUserId() && !User.IsInRole("Admin")
+                ? Unauthorized()
+                : Ok(result);
     }
 
     [HttpPost]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> AddAsync(CreateReservationDto reservation)
     {
+        if(reservation.CustomerId != GetCurrentUserId() && !User.IsInRole("Admin"))
+            return Unauthorized();
+
         var result = await _reservationService.Reserve(reservation);
         return result.Success
             ? Ok(result)
@@ -42,8 +55,11 @@ public class ReservationController : ControllerBase
     }
 
     [HttpPut]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> Update(UpdateReservationDto reservation)
     {
+        if (reservation.CustomerId != GetCurrentUserId() && !User.IsInRole("Admin"))
+            return Unauthorized();
         var result = await _reservationService.Update(reservation);
         return result.Success
             ? Ok(result)
@@ -51,8 +67,13 @@ public class ReservationController : ControllerBase
     }
 
     [HttpDelete]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> Remove(RemoveReservationDto reservation)
     {
+        var reservationCustomerId= _reservationService.GetByIdAsync((int)reservation.Id!).Result.Data!.CustomerId;
+        if (reservationCustomerId != GetCurrentUserId() && !User.IsInRole("Admin"))
+            return Unauthorized();
+
         var result = await _reservationService.Remove(reservation);
         return result.Success
             ? Ok(result)
@@ -60,24 +81,21 @@ public class ReservationController : ControllerBase
     }
 
     [HttpDelete("{id}")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> RemoveById(int id)
     {
+        var reservationCustomerId = _reservationService.GetByIdAsync(id).Result.Data!.CustomerId;
+        if (reservationCustomerId != GetCurrentUserId() && !User.IsInRole("Admin"))
+            return Unauthorized();
+
         var result = await _reservationService.RemoveById(id);
         return result.Success
             ? Ok(result)
             : BadRequest(result.Message);
     }
 
-    [HttpDelete("removerange")]
-    public async Task<IActionResult> RemoveRange(List<RemoveReservationDto> reservations)
-    {
-        var result = await _reservationService.RemoveRange(reservations);
-        return result.Success
-            ? Ok(result)
-            : BadRequest(result.Message);
-    }
-
     [HttpGet("pagination")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin")]
     public IActionResult GetAllPagination([FromQuery] BasePaginationRequest req)
     {
         var result = _reservationService.GetAllPagination(req);
@@ -86,9 +104,14 @@ public class ReservationController : ControllerBase
             : BadRequest(result.Message);
     }
     [HttpGet("cancel-reservation")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Admin, Customer")]
     public async Task<IActionResult> CancelReservation(int reservationId)
     {
-        var result =await _reservationService.CancelReservationById(reservationId);
+        var reservationCustomerId = _reservationService.GetByIdAsync(reservationId).Result.Data!.CustomerId;
+        if (reservationCustomerId != GetCurrentUserId() && !User.IsInRole("Admin"))
+            return Unauthorized();
+
+        var result = await _reservationService.CancelReservationById(reservationId);
         return result.Success
             ? Ok(result)
             : BadRequest(result.Message);
